requirement.txt
asn1crypto==0.24.0
backcall==0.1.0
braintree==3.54.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.3.2
dialogflow==0.5.2
Django==2.1.5
django-bootstrap-datepicker-plus==3.0.5
django-extensions==2.1.5
django-payments==0.13.0
django-registration-redux==2.6
google-api-core==1.7.0
google-auth==1.6.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
protobuf==3.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
stripe==2.29.4
suds-jurko==0.6
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
xmltodict==0.12.0
